# bismarck-game/backend/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    name: Bismarck Game Team
  description: API для игры Bismarck Chase - пошаговой стратегической игры о морских
    сражениях
  title: Bismarck Game API
  version: "1.0"
paths:
...
  /ships/all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получение всех кораблей
      tags:
      - Ships
  /ships/config/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID корабля
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties: true
            type: object
      summary: Получение конфигурации корабля
      tags:
      - Ships
  /ships/create-unit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для создания юнита
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Создание юнита из конфигурации
      tags:
      - Ships
  /ships/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: Сторона
        in: query
        name: side
        type: string
      - description: Тип корабля
        in: query
        name: type
        type: string
      - description: Минимальное топливо
        in: query
        name: min_fuel
        type: integer
      - description: Максимальное топливо
        in: query
        name: max_fuel
        type: integer
      - description: Минимальное уклонение
        in: query
        name: min_evasion
        type: integer
      - description: Максимальное уклонение
        in: query
        name: max_evasion
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Поиск кораблей по критериям
      tags:
      - Ships
  /ships/side/{side}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Сторона (german/allied)
        in: path
        name: side
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получение доступных кораблей для стороны
      tags:
      - Ships
  /ships/stats:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получение статистики конфигурации
      tags:
      - Ships
  /ships/type/{type}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Тип корабля
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получение кораблей по типу
      tags:
      - Ships
  /ships/types:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получение типов кораблей
      tags:
      - Ships
securityDefinitions:
  Bearer:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"